DROID-2198 fixes
