+++ app/src/main/java/com/anytypeio/anytype/di/feature/onboarding/signup/OnboardingMnemonicDI.kt
@@ -5,6 +5,7 @@ import com.anytypeio.anytype.analytics.base.Analytics
 import com.anytypeio.anytype.di.common.ComponentDependencies
 import com.anytypeio.anytype.domain.auth.interactor.GetMnemonic
 import com.anytypeio.anytype.domain.auth.repo.AuthRepository
+import com.anytypeio.anytype.domain.config.ConfigStorage
 import com.anytypeio.anytype.presentation.onboarding.signup.OnboardingMnemonicViewModel
 import dagger.Binds
 import dagger.Component
@@ -55,6 +56,7 @@ object OnboardingMnemonicModule {
 interface OnboardingMnemonicDependencies : ComponentDependencies {
     fun authRepository(): AuthRepository
     fun analytics(): Analytics
+    fun config(): ConfigStorage
 }
 
 @Scope
