+++ presentation/src/main/java/com/anytypeio/anytype/presentation/onboarding/OnboardingStartViewModel.kt
@@ -5,9 +5,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.anytypeio.anytype.analytics.base.Analytics
 import com.anytypeio.anytype.analytics.base.EventsDictionary
-import com.anytypeio.anytype.analytics.base.EventsDictionary.ScreenOnboardingStep.VOID
 import com.anytypeio.anytype.analytics.base.sendEvent
-import com.anytypeio.anytype.presentation.extension.sendAnalyticsOnboardingScreenEvent
 import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
@@ -50,10 +48,6 @@ class OnboardingStartViewModel @Inject constructor(
         viewModelScope.launch {
             navigation.emit(destination)
         }
-    }
-
-    fun sendAnalyticsOnboardingScreen() {
-        viewModelScope.sendAnalyticsOnboardingScreenEvent(analytics, VOID)
     }
 
     interface AuthNavigation {
