+++ presentation/src/main/java/com/anytypeio/anytype/presentation/onboarding/signup/OnboardingMnemonicViewModel.kt
@@ -6,8 +6,10 @@ import androidx.lifecycle.viewModelScope
 import com.anytypeio.anytype.analytics.base.Analytics
 import com.anytypeio.anytype.analytics.base.EventsDictionary
 import com.anytypeio.anytype.domain.auth.interactor.GetMnemonic
+import com.anytypeio.anytype.domain.config.ConfigStorage
 import com.anytypeio.anytype.presentation.extension.sendAnalyticsOnboardingClickEvent
 import com.anytypeio.anytype.presentation.extension.sendAnalyticsOnboardingScreenEvent
+import com.anytypeio.anytype.presentation.extension.sendOpenAccountEvent
 import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
@@ -15,12 +17,16 @@ import timber.log.Timber
 
 class OnboardingMnemonicViewModel @Inject constructor(
     private val getMnemonic: GetMnemonic,
-    private val analytics: Analytics
+    private val analytics: Analytics,
+    private val configStorage: ConfigStorage
 ) : ViewModel() {
 
     val state = MutableStateFlow<State>(State.Idle(""))
 
     init {
+        viewModelScope.sendAnalyticsOnboardingScreenEvent(analytics,
+            EventsDictionary.ScreenOnboardingStep.PHRASE
+        )
         viewModelScope.launch {
             proceedWithMnemonicPhrase()
         }
@@ -43,6 +49,14 @@ class OnboardingMnemonicViewModel @Inject constructor(
             type = EventsDictionary.ClickOnboardingButton.CHECK_LATER,
             step = EventsDictionary.ScreenOnboardingStep.PHRASE
         )
+        viewModelScope.launch {
+            val config = configStorage.getOrNull()
+            if (config != null) {
+                analytics.sendOpenAccountEvent(
+                    analytics = config.analytics
+                )
+            }
+        }
     }
 
     private suspend fun proceedWithMnemonicPhrase() {
@@ -52,12 +66,6 @@ class OnboardingMnemonicViewModel @Inject constructor(
                 state.value = State.Mnemonic(mnemonic)
                 Any()
             }
-        )
-    }
-
-    fun sendAnalyticsOnboardingScreen() {
-        viewModelScope.sendAnalyticsOnboardingScreenEvent(analytics,
-            EventsDictionary.ScreenOnboardingStep.SOUL
         )
     }
 
@@ -72,13 +80,15 @@ class OnboardingMnemonicViewModel @Inject constructor(
 
     class Factory @Inject constructor(
         private val getMnemonic: GetMnemonic,
-        private val analytics: Analytics
+        private val analytics: Analytics,
+        private val configStorage: ConfigStorage,
     ) : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
             return OnboardingMnemonicViewModel(
                 getMnemonic = getMnemonic,
-                analytics = analytics
+                analytics = analytics,
+                configStorage = configStorage
             ) as T
         }
     }
